Refactor unnecessary code that has since been replaced; fix logic determining who won; fix pathing to files
